Separate out the document helpers from the media helpers
