+++ src/wagtail_unveil/management/commands/list_admin_urls.py
@@ -9,7 +9,7 @@ import re
 from urllib.parse import urljoin
 
 from wagtail_unveil.helpers.image_helpers import get_image_admin_urls
-from wagtail_unveil.helpers.media_helpers import get_document_admin_urls
+from wagtail_unveil.helpers.document_helpers import get_document_admin_urls
 from wagtail_unveil.helpers.modeladmin_helpers import (
     get_modeladmin_models,
     get_modeladmin_urls,
